Start omskrivning for felles eventlager
