+++ src/main/kotlin/no/nav/k9/los/nyoppgavestyring/domeneadaptere/k9/eventmottak/eventlager/EventlagerKonverteringsjobb.kt
@@ -1,6 +1,7 @@
-package no.nav.k9.los.nyoppgavestyring.domeneadaptere.k9.eventmottak.punsj
+package no.nav.k9.los.nyoppgavestyring.domeneadaptere.k9.eventmottak.eventlager
 
 import kotliquery.queryOf
+import no.nav.k9.los.nyoppgavestyring.domeneadaptere.k9.eventmottak.punsj.K9PunsjEventRepository
 import no.nav.k9.los.nyoppgavestyring.infrastruktur.db.TransactionalManager
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
@@ -9,12 +10,12 @@ import java.time.temporal.ChronoUnit
 import java.util.UUID
 import kotlin.concurrent.thread
 
-class EventlagerKonverteringsjobbb(
+class EventlagerKonverteringsjobb(
     private val transactionalManager: TransactionalManager,
     private val punsjEventRepository: K9PunsjEventRepository,
     private val eventlagerKonverteringsservice: EventlagerKonverteringsservice,
 ) {
-    private val log: Logger = LoggerFactory.getLogger(EventlagerKonverteringsjobbb::class.java)
+    private val log: Logger = LoggerFactory.getLogger(EventlagerKonverteringsjobb::class.java)
     private val TRÅDNAVN = "eventlagerKonverteringPunsj"
 
     fun kjørEventlagerKonvertering() {
