+++ src/main/kotlin/no/nav/k9/los/nyoppgavestyring/domeneadaptere/k9/eventmottak/punsj/K9PunsjEventHandler.kt
@@ -12,6 +12,7 @@ import no.nav.k9.los.domene.repository.StatistikkRepository
 import no.nav.k9.los.nyoppgavestyring.domeneadaptere.k9.eventmottak.EventHandlerMetrics
 import no.nav.k9.los.nyoppgavestyring.domeneadaptere.k9.eventmottak.EventTeller
 import no.nav.k9.los.nyoppgavestyring.domeneadaptere.k9.eventmottak.IModell
+import no.nav.k9.los.nyoppgavestyring.domeneadaptere.k9.eventmottak.eventlager.EventlagerKonverteringsservice
 import no.nav.k9.los.nyoppgavestyring.domeneadaptere.k9.eventtiloppgave.punsjtillos.K9PunsjTilLosAdapterTjeneste
 import no.nav.k9.los.nyoppgavestyring.infrastruktur.azuregraph.IAzureGraphService
 import no.nav.k9.los.nyoppgavestyring.infrastruktur.db.TransactionalManager
