+++ src/main/resources/migreringer/V1.0_0083__ny_eventhistorikk_punsj.sql
@@ -1,14 +1,15 @@
-create table if not exists eventlager_punsj
+create table if not exists eventlager
 (
     id                          BIGINT GENERATED ALWAYS AS IDENTITY     NOT NULL PRIMARY KEY,
+    FAGSYSTEM                   VARCHAR(50)     NOT NULL,
     EKSTERN_ID                  VARCHAR(100)    NOT NULL,
     EKSTERN_VERSJON             VARCHAR(100)    NOT NULL,
     INTERN_VERSJON              INT             NOT NULL,
     DATA                        jsonb           NOT NULL,
     dirty                       boolean         NOT NULL DEFAULT TRUE,
     opprettet                   timestamp       NOT NULL default localtimestamp,
-    unique (EKSTERN_ID, EKSTERN_VERSJON),
-    unique (EKSTERN_ID, INTERN_VERSJON)
+    unique (FAGSYSTEM, EKSTERN_ID, EKSTERN_VERSJON),
+    unique (FAGSYSTEM, EKSTERN_ID, INTERN_VERSJON)
 );
 
 CREATE TABLE if not exists eventlager_punsj_historikkvask_ferdig
